requirements.txt
functions-framework==3.0.0
google-cloud-storage==2.14.0
google-cloud-bigquery==3.21.0
google-cloud-aiplatform==1.49.0


